refactor: 검증 로직 분리
